Moving target_list function to core.util
