+++ pyomo/gdp/plugins/bigm.py
@@ -26,12 +26,13 @@ from pyomo.core.base.external import ExternalFunction
 from pyomo.core.base import Transformation, TransformationFactory, Reference
 import pyomo.core.expr.current as EXPR
 from pyomo.gdp import Disjunct, Disjunction, GDP_Error
-from pyomo.gdp.util import ( _warn_for_active_logical_constraint, target_list,
-                             is_child_of, get_src_disjunction,
-                             get_src_constraint, get_transformed_constraints,
+from pyomo.gdp.util import ( _warn_for_active_logical_constraint, is_child_of,
+                             get_src_disjunction, get_src_constraint,
+                             get_transformed_constraints,
                              _get_constraint_transBlock, get_src_disjunct,
                              _warn_for_active_disjunction,
                              _warn_for_active_disjunct, preprocess_targets)
+from pyomo.core.util import target_list
 from pyomo.network import Port
 from pyomo.repn import generate_standard_repn
 from functools import wraps
